requirements.txt
ckanapi-exporter
